e2e_tests/pyproject.toml
[tool.poetry]
name = "e2e-tests"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.1,<4.0"
pytest-playwright-visual = "^2.1.2"
pytest-playwright = "^0.7.0"
pytest-html = "^4.1.1"
PyJWT = "^2.8.0"
itsdangerous = "^2.2.0"
python-keycloak = "^5.8.1"
playwright = "^1.55.0"
cryptography = "45.0.6"
numpy = "^1.24.0"
scikit-image = "^0.25.2"




[tool.pytest.ini_options]
addopts = [
    "--html=playwright-report/report.html",
    "--self-contained-html",
    "--screenshot=only-on-failure",
    "--video=retain-on-failure",
    "--output=test-results",
    "--tracing=retain-on-failure"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
